[Benchmark] Move MMMU answer parsers to metrics/accuracy
First slice of the metric-consolidation refactor proposed in #360.

Move parse_multi_choice_response, parse_open_response, and eval_open
(plus their private number-normalising helpers _check_is_number,
_normalize_str, _extract_numbers, _parse_open_answer_tag) from
benchmarks/tasks/visual_understand.py into benchmarks/metrics/accuracy.py.
Function bodies are copied verbatim; no behavioural change.

- benchmarks/tasks/visual_understand.py keeps re-exports of the three
  public helpers as a one-release shim.
- benchmarks/tasks/video_understanding.py imports the parser from
  metrics.accuracy directly.
- tests/test_metrics_accuracy.py adds 27 unit tests pinning the
  parsing behaviour (extract_answer_letter table, multi-choice
  bracket/space/option-text/fallback paths, open-answer tag/boxed/
  number extraction, eval_open string and numeric matching).

Refs: #360
