+++ benchmarks/tasks/visual_understand.py
@@ -1,5 +1,11 @@
 # SPDX-License-Identifier: Apache-2.0
-"""MMMU benchmark case -- answer parsing, send_fn, metrics, and persistence."""
+"""MMMU benchmark case -- send_fn, metrics, and persistence.
+
+Answer-parsing helpers (``parse_multi_choice_response``,
+``parse_open_response``, ``eval_open``) live in
+``benchmarks.metrics.accuracy``. They are re-exported from this module
+for backwards compatibility with existing call sites.
+"""
 
 from __future__ import annotations
 
@@ -8,7 +14,6 @@ import base64
 import logging
 import os
 import random
-import re
 import time
 
 import aiohttp
@@ -16,6 +21,21 @@ import aiohttp
 from benchmarks.benchmarker.data import RequestResult
 from benchmarks.benchmarker.runner import SendFn
 from benchmarks.dataset.mmmu import MMMUSample, image_to_data_uri
+from benchmarks.metrics.accuracy import (
+    eval_open,
+    parse_multi_choice_response,
+    parse_open_response,
+)
+
+__all__ = [
+    "MULTI_CHOICE_INSTRUCTION",
+    "compute_mmmu_metrics",
+    "eval_open",
+    "make_mmmu_send_fn",
+    "parse_multi_choice_response",
+    "parse_open_response",
+    "print_mmmu_accuracy_summary",
+]
 
 logger = logging.getLogger(__name__)
 
@@ -29,180 +49,6 @@ MULTI_CHOICE_INSTRUCTION = (
     "where LETTER is one of the options. "
     "Think step by step before answering."
 )
-
-
-def _check_is_number(s: str) -> bool:
-    try:
-        float(s.replace(",", ""))
-        return True
-    except ValueError:
-        return False
-
-
-def _normalize_str(s: str) -> list[float | str]:
-    """Normalize a string for open-ended answer comparison."""
-    s = s.strip()
-    if _check_is_number(s):
-        return [round(float(s.replace(",", "")), 2)]
-    return [s.lower()] if len(s) > 1 else [" " + s, s + " "]
-
-
-def _extract_numbers(s: str) -> list[str]:
-    """Extract all numbers (with commas, scientific notation, etc.) from *s*."""
-    pattern_commas = r"-?\b\d{1,3}(?:,\d{3})+\b"
-    pattern_scientific = r"-?\d+(?:\.\d+)?[eE][+-]?\d+"
-    pattern_simple = r"-?(?:\d+\.\d+|\.\d+|\d+\b)(?![eE][+-]?\d+)(?![,\d])"
-    return (
-        re.findall(pattern_commas, s)
-        + re.findall(pattern_scientific, s)
-        + re.findall(pattern_simple, s)
-    )
-
-
-def _parse_open_answer_tag(response: str) -> str | None:
-    """Try to extract the answer from an explicit 'Answer: ...' line.
-
-    Supports formats like ``Answer: 42``, ``Answer: MgS``,
-    ``Answer: \\boxed{13.0}``.  Returns ``None`` when no match is found.
-    """
-    matches = re.findall(
-        r"[Aa]nswer\s*:\s*\*?\*?\s*(.+)",
-        response,
-    )
-    if not matches:
-        return None
-    raw = matches[-1].strip().rstrip(".")
-    # Unwrap \boxed{...} if present
-    boxed = re.search(r"\\boxed\{(.+?)\}", raw)
-    if boxed:
-        raw = boxed.group(1)
-    # Strip surrounding ** (bold markdown)
-    raw = raw.strip("*").strip()
-    return raw if raw else None
-
-
-def parse_open_response(response: str) -> list[float | str]:
-    """Extract answer candidates from an open-ended model response.
-
-    First tries to extract from an explicit ``Answer: ...`` line.
-    Falls back to heuristic key-subresponse extraction.
-    """
-    # Fast path: explicit "Answer: ..."
-    tag_answer = _parse_open_answer_tag(response)
-    if tag_answer is not None:
-        out: list = []
-        out.extend(_normalize_str(tag_answer))
-        for num in _extract_numbers(tag_answer):
-            out.extend(_normalize_str(num))
-        return list(dict.fromkeys(out))
-
-    # Fallback: heuristic extraction
-    def _get_key_subresponses(resp: str) -> list[str]:
-        resp = resp.strip().strip(".").lower()
-        subs = re.split(r"\.\s(?=[A-Z])|\n", resp)
-        indicators = [
-            "could be ",
-            "so ",
-            "is ",
-            "thus ",
-            "therefore ",
-            "final ",
-            "answer ",
-            "result ",
-        ]
-        keys: list[str] = []
-        for i, s in enumerate(subs):
-            cands = indicators + ["="] if i == len(subs) - 1 else indicators
-            shortest = None
-            for ind in cands:
-                if ind in s:
-                    part = s.split(ind)[-1].strip()
-                    if not shortest or len(part) < len(shortest):
-                        shortest = part
-            if shortest and shortest not in (":", ",", ".", "!", "?", ";", "'"):
-                keys.append(shortest)
-        return keys or [resp]
-
-    key_resps = _get_key_subresponses(response)
-    pred_list = key_resps.copy()
-    for r in key_resps:
-        pred_list.extend(_extract_numbers(r))
-    out = []
-    for x in pred_list:
-        out.extend(_normalize_str(x))
-    return list(dict.fromkeys(out))
-
-
-def eval_open(gold: str | list[str], preds: list[float | str]) -> bool:
-    """Check if any prediction matches the gold answer (fuzzy)."""
-    if isinstance(gold, list):
-        norm_answers: list = []
-        for ans in gold:
-            norm_answers.extend(_normalize_str(ans))
-    else:
-        norm_answers = _normalize_str(gold)
-    for p in preds:
-        if isinstance(p, str):
-            for na in norm_answers:
-                if isinstance(na, str) and na in p:
-                    return True
-        else:
-            if p in norm_answers:
-                return True
-    return False
-
-
-def parse_multi_choice_response(
-    response: str,
-    all_choices: list[str],
-    index2ans: dict[str, str],
-) -> tuple[str, bool]:
-    """Extract a single answer letter from the model response.
-
-    Priority: ``Answer: X`` → ``(A)`` bracket → ``·A·`` space-padded →
-    option-text match → last-occurrence tie-break → random fallback.
-
-    Returns ``(choice, is_fallback)``. ``is_fallback`` is ``True`` iff
-    nothing could be parsed out of *response* and a random choice was
-    returned — this counter is observational and doesn't change scoring
-    behavior vs. the MMMU reference eval.
-    """
-    answer_matches = re.findall(r"[Aa]nswer\s*:\s*\*?\*?\s*\(?([A-Z])\)?", response)
-    if answer_matches:
-        candidate = answer_matches[-1]
-        if candidate in all_choices:
-            return candidate, False
-
-    for char in (",", ".", "!", "?", ";", ":", "'"):
-        response = response.strip(char)
-    response = " " + response + " "
-
-    candidates: list[str] = []
-    for choice in all_choices:
-        if f"({choice})" in response:
-            candidates.append(choice)
-    if not candidates:
-        for choice in all_choices:
-            if f" {choice} " in response:
-                candidates.append(choice)
-    if not candidates and len(response.split()) > 5:
-        for idx, ans in index2ans.items():
-            if ans and ans.lower() in response.lower():
-                candidates.append(idx)
-    if not candidates:
-        return random.choice(all_choices), True
-    if len(candidates) == 1:
-        return candidates[0], False
-
-    starts: list[int] = []
-    for can in candidates:
-        pos = response.rfind(f"({can})")
-        if pos == -1:
-            pos = response.rfind(f" {can} ")
-        if pos == -1 and index2ans.get(can):
-            pos = response.lower().rfind(index2ans[can].lower())
-        starts.append(pos)
-    return candidates[max(range(len(candidates)), key=starts.__getitem__)], False
 
 
 def make_mmmu_send_fn(
